refactor(iot): defer the connector base to its own PR (OJD-1668)
This PR only needs the trust validator and family types from @repo/iot;
the DeviceDriver template refactor, GenericCommandConnector, AmbitConnector,
identifyDevice, and their specs move to a stacked follow-up so the diff here
stays consolidation-only. A small core/progress.ts keeps the CommandProgress
shape the workbook ports reference; the connector base will emit it natively.
