+++ apps/web/hooks/workbook/useWorkbookExecution/useWorkbookExecution.ts
@@ -114,7 +114,8 @@ export function useWorkbookExecution({
             throw new Error("No device connected - connect a device to run this protocol");
           }
           const onAbort = () => {
-            void driverRef.current?.cancel?.();
+            // Cancel is a MultispeQ driver capability, not yet on IDeviceDriver (OJD-1668).
+            void (driverRef.current as { cancel?: () => Promise<void> } | null)?.cancel?.();
           };
           signal.addEventListener("abort", onAbort);
           try {
