+++ packages/iot/tsconfig.build.json
@@ -6,5 +6,5 @@
     "declaration": true,
     "declarationMap": true
   },
-  "exclude": ["src/**/*.spec.ts", "src/**/testing/**"]
+  "exclude": ["src/**/*.spec.ts"]
 }
